// src/generated/models/Ale_naviosesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ale_naviosesstatecode = {
  0: 'Ativo(a)',
  1: 'Inativo(a)'
} as const;
export type Ale_naviosesstatecode = keyof typeof Ale_naviosesstatecode;
export const Ale_naviosesstatuscode = {
  1: 'Ativo(a)',
  2: 'Inativo(a)'
} as const;
export type Ale_naviosesstatuscode = keyof typeof Ale_naviosesstatuscode;

export interface Ale_naviosesBase {
  ale_latitude?: string;
  ale_longitude?: string;
  ale_mmsi?: string;
  ale_naviosid: string;
  ale_nomedonavio: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ale_naviosesstatecode;
  statuscode?: Ale_naviosesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ale_navioses extends Ale_naviosesBase {
  createdon?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}